Logical cycle of Create, Read, Update and Delete operations in a binary search tree
The Java project provides a menu driven interface that allows a user to
exercise the aforementioned operations on a binary search tree of
integers.

Implements a class called BinarySearchTree for representing binary
search trees of integers. The class supports the following operations:

1.creation - allows a user to create an empty binary search tree.
2. insertion - allows a user to insert a single integer into the tree.
3. contains - allows a user to check if an integer belongs to the tree.
4. delete - allows a user to delete a single integer from the tree.

In addition, the class supports the ability to print the contents of
the tree using a preorder, postorder, or inorder traversal.